requirements.txt
flask
flask-migrate
flask-sqlalchemy
gunicorn
psycopg2
pygithub
pymessenger
pymysql
shortuuid
udacity
